AI_features/item_price_advisor/requirements.txt
langchain
langchain-google-genai
google-search-results
python-dotenv
google-api-python-client